- change most of the ids from uint32_t to int32_t. I wonder what hit me when I changed them all to unsigned ... - remove the compiler warnings in the server

+++ server/main.c
@@ -92,7 +92,7 @@ int rc_init(void)
 		 * and all his sockets up */
 		while (!running)  
 			nanosleep(&t,NULL); 	
-		for (i=0;i<(sizeof(s3drc)/sizeof(char **));i++)
+		for (i=0 ; i < ((int)(sizeof(s3drc)/sizeof(char **))) ; i++)
 		{
 			if ((*s3drc[i])!=NULL)
 			{
